+++ src/main/java/com/spectrum/service/post/PostService.java
@@ -1,4 +1,4 @@
-package com.spectrum.service;
+package com.spectrum.service.post;
 
 import com.spectrum.controller.post.dto.PostCreateResponse;
 import com.spectrum.controller.post.dto.PostDetailResponse;
@@ -11,9 +11,9 @@ import com.spectrum.exception.user.UserNotFoundException;
 import com.spectrum.exception.post.PostNotFoundException;
 import com.spectrum.repository.post.PostRepository;
 import com.spectrum.repository.user.UserRepository;
-import com.spectrum.service.dto.PostCreateDto;
-import com.spectrum.service.dto.PostDto;
-import com.spectrum.service.dto.PostUpdateDto;
+import com.spectrum.service.post.dto.PostCreateDto;
+import com.spectrum.service.post.dto.PostDto;
+import com.spectrum.service.post.dto.PostUpdateDto;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
