+++ src/main/java/com/spectrum/service/post/dto/PostCreateDto.java
@@ -1,4 +1,4 @@
-package com.spectrum.service.dto;
+package com.spectrum.service.post.dto;
 
 import com.spectrum.domain.post.Post;
 import lombok.Getter;
